requirements.txt
pyyaml==6.0
aiofiles==23.1.0
pydantic==1.10.7
pydantic[dotenv]==1.10.7
fastapi==0.95.2
orjson==3.8.12